Updated the FOV to 90, added an Audiosource to the Player
The default FOV was 60 which is rather cramped for PC.
The AudioSource is by default assigned with the pop sound, it gets played from the BoostJumpHeight script. (Decided to not have the emerald have the audio source since it wouldn't play because it gets destroyed, would be cumbersome to build around it).